Modified rotrel to match Bird 2013 (3.32)
It was brought to our attention by Maninder Grover and Dr. Paolo
Valentini that including the rotational energy of the relaxing particle
in the call to rotrel resolved the issue with temperature varying
rotational relaxation not maintaining equilibrium at steady state.

Eq (3.32) in Bird 2013 discusses the same inclusion, with the added
detail of including the rotational DoF in the calculation for T_coll.
There are discussions of shortcomings of this approach related to
partially excited modes, but we feel the steady state equilibrium issue
is much more significant.

This does deviate from the original Parker (1959) derivation of the
temperature dependence of Zrot, but our tests examining both the
relaxation rate and the rotational energy distribution function show no
significant deviations from what is expected.

Further analysis is being performed to understand exactly why the
non-equilibrium is not maintained with the original implementation and
to hopefully provide insights to fixes of the issues with the current
implementation discussed by Bird.