First pass at cell T variable relaxation number
Added logic so that the compute to calculate a per cell temperature is
done at the start of the collide algorithm each timestep. Then, it pulls
the temperature from this compute for whatever grid cell it is in while
doing collisions.

The compute used in a hard-coded "coll_temp" as the name of the compute.
Since this compute does not exist in any of the examples, SPARTA reverts
to the old behavior and all the existing tests pass. The new
functionality is currently untested.