Added variable collision temp example
Added an example that uses a variable to calculate the weighted average of
translational and rotational temperatures to use as the collision temperature
for relaxation.

Began stubbing out the logic for this, by handling the c_ and v_ prefixes
in the name stored to calculate the temperature and removed the _compute
name from the input to make it more general.