refactor: Move electronic mode data into struct
Created a dedicated struct to hold the data describing the electronic
mode, shrinking the Species struct back to (almost) its original size
and enabling only species with electronic modes defined in the input to
allocate the memory required to hold them. This also removes the limits
of the MAXSTATES and MAXSPECIES constants from the data structure sizes,
although these constants are still used in parsing.