refactor: Dynamically allocated ElecElec struct
Changed the members of the ElecFile struct from fixed length arrays/
vectors to dynamically allocated. This removed the compile time limits
on the number of states and species in the simulation for this part of
parsing and should reduce the memory used during parsing. This parallels
changes in the previous commits to make the ElectronicData struct used
during the simulations dynamically allocated.