fix: Resolved inconsistent data in fix_elecmode::update_custom
Added a new method, particle::ielec, that samples the index of an
electronic state from a boltzmann distribution and simply use the result
of that for particle::eelec. This resolves the issue that eelec is much
more widely used but going from eelec to ielec is very difficult while
the other direction is straightforward.