fix: Eliminate infinite loop in electronic temp calc
In highly excited cells, it was possible to get -inf as the first guess
for the bisection method when calculation the electronic temperature.
The bisection method implemented there would then infinitely loop,
hanging the simulation. This commit simply checks for the -inf value and
uses a different but sane value for the initial guess.