fix: Change initial guess for T_elec calc
It was found that the previous initial guess logic for calculating the
electronic temperature often failed to converge due to the strange
behavior of the electronic temperature at relatively high excitation
levels where the temperature rolls around from positive infinity to
negative infinity. It was found via experimentation that simply using
the temperature of the first excited state for the species of interest
was much more robust.