feat: Changed elec dof to input from calc
It was found that the calculation of the electronic degrees of freedom
from the energy, using only a single particle, caused non-physical
degrees of freedom to be used in the TCE model. Specifically, when a
particle was highly excited (usually in the highest 1 or 2 levels for
its species), the calculation would lead to a negative value. This
caused the chemistry rates to be incorrect.

After a search of the literature I was unable to find a method that
others had used that didn't depend on cell averages. This change adds
another input to the elec file where the user can set the value. Small
values have a small effect on chemistry, but this allows the reactivity
of specific states to be tuneds as well.