fix: Enable reread of surfs in sims with particles
If a surface was read after particles have been created and the `particle
keep` flag was turned on to leave all particles in place, the code would
throw an error. It turns out, if any grid cells were completely inside
the surface then the error was produced, even if no particles were
inside that cell. This commit simply adds a condition to the error that
there must be particles inside the cell.