Update surf2paraview.py
This fixes #464. 

While #499 was intended to fix this issue, it introduced a new bug. The bug is that the type id was read as one of the point ids. This is because the type is written to the 2nd field.

The present pull request amends this by reading the point ids in reverse order from the end of the line.

Note: while this now supports type, it will fail catastrophically if write_surf is used with custom arg. There is no adhoc solution to support the general case because it would involve unknown number of fields (1 or 2 with type) before the point ids and unknown number of fields after.